refactor: use builder instead of pre-created page
By creating a list with a `page: Page()` parameter, the `Page()` widget is being instantiated and built with all of their logic, data, etc (not the render part). The problem here is that three whole pages are being instantiated and no being used. Imagine this happening to the location page where you'll have maps controllers, and all the other things.
To mitigate this early creation, in this case where the page is delivered to a navigator builder, we use the builder function as the parameter instead of only the page widget.

Inside TournamentListItem, before:

```
MaterialPageRoute(
  builder: (context) => PARAMETER,
  settings: RouteSettings(
    name: widget.route,
  ),
),
```

after:

```
MaterialPageRoute(
  builder: PARAMETER,
  settings: RouteSettings(
    name: widget.route,
  ),
),
```